requirements.txt
streamlit
azure-identity
azure-mgmt-subscription
azure-mgmt-resource
azure-mgmt-managementgroups
